backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
dateparser==1.2.2
exceptiongroup==1.3.1
fastapi==0.128.8
frozenlist==1.8.0
h11==0.16.0
idna==3.11
multidict==6.7.1
propcache==0.4.1
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-binance==1.0.34
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2026.1.15
requests==2.32.5
six==1.17.0
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.39.0
websockets==15.0.1
yarl==1.22.0
pandas
